requirements.txt
datefinder==0.7.0
Flask==1.1.1
gunicorn==20.0.0
itsdangerous==1.1.0
Jinja2==2.10.1
os-sys-php==2019.10.13
Pillow==6.2.1
pytesseract==0.3.0
regex==2019.11.1
requests==2.21.0
requests-download==0.1.2
Werkzeug==0.16.0
wget==3.2

